% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{Annotate}
\alias{Annotate}
\title{Annotate raw data ~~under construction~~}
\usage{
Annotate(path, layout.blank.df, filename, out.dir)
}
\arguments{
\item{path}{path to edited plate data. This folder should be automatically created with the GCG::Import function. this folder must contain only the files you wish to annotate.}

\item{layout.blank.df}{dataframe containing your combined layout and blank information}

\item{filename}{what would you like to name the output file? please include .xlsx}

\item{out.dir}{where would you like to save the file?}
}
\description{
Step through raw data and add strain and condition information. Each plate exported as a sheet in an Excel file
}
