% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAUC.R
\name{eAUC}
\alias{eAUC}
\title{eAUC -- Emperical area under the curve}
\usage{
eAUC(time.vector, y.vector)
}
\arguments{
\item{time.vector}{vectored list of time values}

\item{y.vector}{a list of y value datapoints corresponding to the intervals of your time values.}
}
\value{
Unit of output is the same as the unit of your y axis. for example, eAUC of an OD600 curve would have a unit "OD600"
}
\description{
Finds the area of the curve of a set of points by integrating every point along your curve as an area value. This function is the core function of the Growthcurver package's EmpericalAreaUnderTheCurve function
}
\details{
Input is two vectors of the same length, one being a "time" vector, and the other your corresponding y value vector
}
