% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelToCSV.R
\name{ExcelToCSV}
\alias{ExcelToCSV}
\title{Convert Excel workbook into separate CSV files}
\usage{
ExcelToCSV(path, out_dir = NULL)
}
\arguments{
\item{path}{is path to a multi-sheet excel file}

\item{out.dir}{out directory: directory where you wish to send these new files, if null, it will send to current directory}
}
\value{
multiple .csv files
}
\description{
Separate Excel worksheets into separate .csv files. filenames will be your.sheetname.csv. it is important to make these sheet names exactly the names of your plates, corresponding to the PLATE names in your plate layout and plate blank files.
}
