% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCGplot.R
\name{GCGplot_bioreps}
\alias{GCGplot_bioreps}
\title{Wrap individual biological replicates to spot-check}
\usage{
GCGplot_bioreps(data.combined.summarized, title, path)
}
\arguments{
\item{data.combined.summarized}{tidy, long data from SummarizeDataCombined}

\item{title}{what you would like to title this graphic}

\item{path}{path to the folder where you would like to store these pictures -- should end in "/"}
}
\description{
Subsets each biological replicate and plots a facet_wrap with every condition. this allows you to spot check the biological reps to see if there is any obvious problems.
}
