% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCGplot.R
\name{GCGplot_conds}
\alias{GCGplot_conds}
\title{Plot growth curves by condition}
\usage{
GCGplot_conds(data.combined.summarized, graphic.title, path)
}
\arguments{
\item{data.combined.summarized}{Summarized data from ReplicateSummary}

\item{graphic.title}{general title for all conditions. individual condition name will be appended.}

\item{path}{path to the folder where you would like to store these pictures -- should end in "/"}
}
\description{
Description
}
