% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCGplot.R
\name{GCGplot_wrap}
\alias{GCGplot_wrap}
\title{Plot summary of all conditions}
\usage{
GCGplot_wrap(data.combined.summarized, graphic.title, out.dir)
}
\arguments{
\item{data.combined.summarized}{tidy, long data from either SummarizeBioReps or SummarizeTechReps}

\item{graphic.title}{what you would like to title this graphic}

\item{out.dir}{output directory: path to the folder where you would like to store these pictures -- should end in "/"}
}
\description{
Uses ggplot2's facet_wrap function to plot all conditions and saves as jpeg and svg. This
}
