% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growthcurver_convert.R
\name{Growthcurver_convert}
\alias{Growthcurver_convert}
\title{This function combines the plate name and coordinates from data.combined to one character string separated by a "\%". "A1\%A1" for example. Therefore, plate name and/or coordinate numbers should not contain "\%"}
\usage{
Growthcurver_convert(data.combined)
}
\arguments{
\item{data.combined}{the data.combined dataframe specifically from TimeseriesLayoutBlank()}
}
\description{
This function combines the plate name and coordinates from data.combined to one character string separated by a "\%". "A1\%A1" for example. Therefore, plate name and/or coordinate numbers should not contain "\%"
}
