% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{Import}
\alias{Import}
\title{Import plate reader data}
\usage{
Import(path, plate.reader.type, read.interval)
}
\arguments{
\item{path}{path to folder containing plate.csv files}

\item{plate.reader.type}{plate reader type. must be exactly "spark" or "magellan"}

\item{read.interval}{time interval, in minutes used in your plate experiment. reads taken every hour = 60}
}
\description{
Parse through multiple plate reader data files and combine into large dataframe.plate reader data files must all be in a folder with no other .csv files. i recommend making a new directory for these files. "Plate reader data" would be a good name for it
}
