% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanImport.R
\name{Magellan_import}
\alias{Magellan_import}
\title{used to import data from magellan, please see example in "Examples"}
\usage{
Magellan_import(in.filename)
}
\arguments{
\item{in.filename}{filename, including path, of file you'd like to import.}
}
\description{
used to import data from magellan, please see example in "Examples"
}
