% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanImport.R
\name{MagellanImport}
\alias{MagellanImport}
\title{Import datafiles with Tecan Magellan format}
\usage{
MagellanImport(in.filename)
}
\arguments{
\item{in.filename}{filename, including path, of file you'd like to import.}
}
\description{
used to import data from magellan, please see example in "Examples". Imports raw data from a .csv file. This function is used within "Import"
}
