% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParsePlateLaytout.R
\name{ParsePlateLayout}
\alias{ParsePlateLayout}
\title{Parse through plate matrix}
\usage{
ParsePlateLayout(inDF)
}
\arguments{
\item{inDF}{dataframe to be parsed from plate matrix format into tidy "long" format}
}
\value{
out_df, a tidy, long style dataframe with your plate information
}
\description{
A function to parse through a plate matrix style table and return a "long" dataframe, with each "well" as a row. this function is used inside of the PlateLayout() and PlateBlank() function.
}
