% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateBlank.R
\name{PlateBlank}
\alias{PlateBlank}
\title{A function used to parse through a matrix style plate layouts and create a tidy, long dataframe of the contained information. plate coordinates are read directly from the plate matrix, if you need to change these to match the plate reader datafile output, you can change them no problem. You may copy and paste the plate template as many times as you like. Separate plates are identified in the cell adjcent to "PLATE" in the template. do not change the "PLATE" cell, the script uses this text to find each copied template input. label your plates exactly the same as your plate.reader datafiles and exactly the same as their labels in the similar "plate layout" file. you should have the same amount of plates in each of these 3 locations.}
\usage{
PlateBlank(path)
}
\arguments{
\item{path}{the path (including filename) to plate.blank datafile, modified from template.}
}
\value{
outDF a tidy, long dataframe with the information provided in your input file.
}
\description{
A function used to parse through a matrix style plate layouts and create a tidy, long dataframe of the contained information. plate coordinates are read directly from the plate matrix, if you need to change these to match the plate reader datafile output, you can change them no problem. You may copy and paste the plate template as many times as you like. Separate plates are identified in the cell adjcent to "PLATE" in the template. do not change the "PLATE" cell, the script uses this text to find each copied template input. label your plates exactly the same as your plate.reader datafiles and exactly the same as their labels in the similar "plate layout" file. you should have the same amount of plates in each of these 3 locations.
}
