% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StdPlot.R
\name{std_growth_curve_plot}
\alias{std_growth_curve_plot}
\title{simple ggplot settings used to plot growth curves with 2 variables (Strain/Condition) over time.}
\usage{
std_growth_curve_plot(inDF, title, Read.interval.in.min)
}
\arguments{
\item{inDF}{tidy, long dataframe from TimeseriesLayoutBlank() function or same format}

\item{title}{umbrella tile of your plots}

\item{Read.interval.in.min}{read interval, in minutes of your plate experiment. one read every hour is Read.interval.in.min = 60}
}
\description{
simple ggplot settings used to plot growth curves with 2 variables (Strain/Condition) over time.
}
