% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplicateSummary.R
\name{SummarizeBioReps}
\alias{SummarizeBioReps}
\title{Summarize biological replicates}
\usage{
SummarizeBioReps(TechRepSummary)
}
\arguments{
\item{TechRepSummary}{table produced with TimeseriesLayoutBlank, or similar long table.}
}
\value{
summarized table of your data, including average and SE for every timepoint. this will be used for graphing with ggplot.
}
\description{
Produces mean and SE of technical and biological replicates. if you have multiple biological replicates, this function will first mean the technical replicates within them, then find SE and mean of those means.
}
