% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeseriesLayoutBlank.R
\name{TimeseriesLayoutBlank}
\alias{TimeseriesLayoutBlank}
\title{Combine information from plate layout and blank reading}
\usage{
TimeseriesLayoutBlank(timepoint.df, layout.blank.df)
}
\arguments{
\item{timepoint.df}{the dataframe you recieved from Import() that contains your long, tidy time series data}

\item{layout.blank.df}{dataframe recieved from CombineLayoutBlank() that contains long, tidy plate layout informatin and back-subtract (blank) values.}
}
\value{
data.combined a very long, tidy df with all of your data, corrected by back-subtracting the blank values.
}
\description{
Adds together your dataframes from CombineLayoutBlank and Import to create a very long, tidy dataframe that can be used for plotting and other analysis.
}
