% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCGplot.R
\name{GCGplot_matrices}
\alias{GCGplot_matrices}
\title{Display curves from all wells}
\usage{
GCGplot_matrices(data.combined.summarized, graphic.title, path)
}
\arguments{
\item{data.combined.summarized}{tidy, long data from either SummarizeBioReps or SummarizeTechReps}

\item{graphic.title}{what you would like to title this graphic}

\item{path}{path to the folder where you would like to store these pictures -- should end in "/"}
}
\description{
Plots each curve individually, plotting each plate as one facet_wrap matrix, in 96 well format. this allows for manual spot-checking of technical replicates and individual wells.
}
